+++ src/modules/client/create-announcement/steps/choose-preferred-style.tsx
@@ -4,11 +4,10 @@ import type { TxKeyPath } from '@/core';
 import { useCustomForm } from '@/core';
 import { StepButtons } from '@/modules/shared';
 import { Text, ToggleCard, View } from '@/shared/components';
+import type { StepperFormProps } from '@/types';
 import type { AnnouncementType } from '@/types/announcement';
 
 import { CreateAnnouncementStepNineSchema } from '../schemas';
-import type { StepperFormProps } from './choose-speciality';
-
 export function ChoosePreferredStyle({
   onHandleBack,
   onHandleNext,
